refactor(probe): complete udp migration with services and cli hooks
